+++ CHANGES.md
@@ -6,10 +6,11 @@
 
 * Context impl refactoring, fixed UB
 
-
 ### Added
 
-* Implemented `Eq` and `PartialEq` for `actix::Addr`
+* Implemented `Eq`, `PartialEq`, and `Hash` for `actix::Addr`
+
+* Implemented `Eq`, `PartialEq`, and `Hash` for `actix::Recipient`
 
 
 ## [0.6.2] (2018-06-xx)